backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
ultralytics==8.0.219
pillow==10.1.0
scikit-learn==1.3.2
numpy==1.24.4
opencv-python-headless==4.7.0.72
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
easyocr>=1.7.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0